The responsability to decide the kind of Composite object to create is delegated to swingintegration.example.Platform.
The situation is not very good on the Linux side: under Gnome 2.20 (Mandriva 2008), the Eclipse runtime using the plugin hangs (compared to the whole gnome manager before).

It looks like the UIManager is changed when drawing the vizPanel.

Indeed, the scrollbars are the same as the Eclipse environments, while previously under Linux, is was the default Metal L&F.
I tried to force the Metal L&F but it does not appear to work.