Surprise!
Felix, it would be nice if you could take a look at the implementation of the IALSTreeContentProvider.
The getBeginLine() and getEndLine() methods that are supposed to indicate the beginning and the end of a declaration is not working properly right now.

The implementations are in fr.univartois.cril.alloyplugin.core