Fixed FileSystem example syntax for Alloy 4.
Removed the text regarding loading the palette. I believe it does not work with Alloy4.
Asked the user to project the solution to FileSystem instead.

It looks like there is a problem in the filesystem example.
The assert statements rmRemovesOneFile are wrong both in FS V and FS VI.

I reported the issue last year (December 12, 2006) to alloy@mit.edu. Felix confirmed the problem and said he would take a look at it. I did not get any answer since then AFAIK. 

For the moment, the comments in the model have been replaced by "DLB:counterexample".