Code folding slightly modified: - works for assertions. - no folding appears for 1 line declaration.
There are still some issues depending of Alloy 4:

Signature pos.y and pos.y2 do not map to the begining and the end of the sig declaration.

Anonymous Facts (declared right after Sig declaration) do not have their pos field properly filled.

Regarding predicates and functions, the body.pos or body.closingBrackets are not correct either.

Since Nicolas is planning a new release of A4E this week, and that A4 RC12 in on the way, it would be nice Felix if you could take a look at those issues.