Allow code completion on Signature names. Quite cool, actually. I guess that the same could be done for predicates and function names.
I had to remove the "this/" prefix from the name of the signature found in the label field.

I am not sure if it is a good idea to do code completion on all the names found in the spec file, since there is no contextual filtering done during the completion, and the number of possible candidates might grow quickly!