http://code.google.com/p/alloy4eclipse/issues/detail?id=81
- Updated A4E to run w/ A4 v. 4.2.0 from MIT's SVN commit 9307
- Reordered fr.univartois.cril.alloyplugin dependencies to resolve class not found errors at eclipse startup.
These errors are subtle but stem from the fact that OSGI starts bundles according to a topological sort of their dependencies; thus, if the plugin has P2,P1 in the manifest but P2 depends on P1, we can get "ClassNotFound" exceptions at startup because P1 might not have started yet. To avoid this, it is important to ensure that the order of plugin dependencies in the manifest is consistent with classloader dependencies among these plugins.